Move featured projects to API v2
