+++ server/__init__.py
@@ -133,6 +133,7 @@ def add_api_endpoints(app):
         ProjectsQueriesNoGeometriesAPI,
         ProjectsQueriesNoTasksAPI,
         ProjectsQueriesAoiAPI,
+        ProjectsQueriesFeaturedAPI,
     )
     from server.api.projects.activities import ProjectsActivitiesAPI
     from server.api.projects.contributions import (
@@ -146,8 +147,9 @@ def add_api_endpoints(app):
     from server.api.projects.actions import (
         ProjectsActionsTransferAPI,
         ProjectsActionsMessageContributorsAPI,
+        ProjectsActionsFeatureAPI,
+        ProjectsActionsUnFeatureAPI,
     )
-    from server.api.projects_apis import FeaturedProjectAPI
 
     # Tasks API import
     from server.api.tasks.resources import (
@@ -269,6 +271,7 @@ def add_api_endpoints(app):
     api.add_resource(
         ProjectsQueriesAoiAPI, "/api/v2/projects/<int:project_id>/queries/aoi/"
     )
+    api.add_resource(ProjectsQueriesFeaturedAPI, "/api/v2/projects/queries/featured")
 
     # Projects' addtional resources
     api.add_resource(
@@ -299,15 +302,11 @@ def add_api_endpoints(app):
         "/api/v2/projects/<int:project_id>/actions/transfer-ownership/",
     )
     api.add_resource(
-        FeaturedProjectAPI,
-        "/api/v1/project/feature",
-        endpoint="feature_projects",
-        methods=["GET"],
+        ProjectsActionsFeatureAPI, "/api/v2/projects/<int:project_id>/actions/feature"
     )
     api.add_resource(
-        FeaturedProjectAPI,
-        "/api/v1/project/feature/<int:project_id>",
-        methods=["POST", "DELETE"],
+        ProjectsActionsUnFeatureAPI,
+        "/api/v2/projects/<int:project_id>/actions/remove-feature",
     )
 
     # Tasks REST endpoint
@@ -325,9 +324,6 @@ def add_api_endpoints(app):
     )
     api.add_resource(
         TasksQueriesAoiAPI, "/api/v2/projects/<int:project_id>/tasks/queries/aoi/"
-    )
-    api.add_resource(
-        MappedTasksByUser, "/api/v1/project/<int:project_id>/mapped-tasks-by-user"
     )
     api.add_resource(
         TasksQueriesOwnLockedAPI,
