+++ server/api/projects/actions.py
@@ -4,6 +4,7 @@ from flask_restful import Resource, request, current_app
 from schematics.exceptions import DataError
 
 from server.models.dtos.message_dto import MessageDTO
+from server.services.project_service import ProjectService, NotFound
 from server.services.project_admin_service import ProjectAdminService
 from server.services.messaging.message_service import MessageService
 from server.services.users.authentication_service import token_auth, tm
@@ -126,3 +127,103 @@ class ProjectsActionsMessageContributorsAPI(Resource):
             error_msg = f"Send message all - unhandled error: {str(e)}"
             current_app.logger.critical(error_msg)
             return {"Error": "Unable to send messages to mappers"}, 500
+
+
+class ProjectsActionsFeatureAPI(Resource):
+    @tm.pm_only(True)
+    @token_auth.login_required
+    def post(self, project_id):
+        """
+        Set a project as featured
+        ---
+        tags:
+            - projects
+        produces:
+            - application/json
+        parameters:
+            - in: header
+              name: Authorization
+              description: Base64 encoded session token
+              required: true
+              type: string
+              default: Token sessionTokenHere==
+            - name: project_id
+              in: path
+              description: The unique project ID
+              required: true
+              type: integer
+              default: 1
+        responses:
+            200:
+                description: Featured projects
+            400:
+                description: Bad request
+            403:
+                description: Forbidden
+            404:
+                description: Project not found
+            500:
+                description: Internal Server Error
+        """
+        try:
+            ProjectService.set_project_as_featured(project_id)
+            return {"Success": True}, 200
+        except NotFound:
+            return {"Error": "Project Not Found"}, 404
+        except ValueError as e:
+            error_msg = f"FeaturedProjects POST: {str(e)}"
+            return {"error": error_msg}, 400
+        except Exception as e:
+            error_msg = f"FeaturedProjects GET - unhandled error: {str(e)}"
+            current_app.logger.critical(error_msg)
+            return {"error": error_msg}, 500
+
+
+class ProjectsActionsUnFeatureAPI(Resource):
+    @tm.pm_only(True)
+    @token_auth.login_required
+    def post(self, project_id):
+        """
+        Unset a project as featured
+        ---
+        tags:
+            - projects
+        produces:
+            - application/json
+        parameters:
+            - in: header
+              name: Authorization
+              description: Base64 encoded session token
+              required: true
+              type: string
+              default: Token sessionTokenHere==
+            - name: project_id
+              in: path
+              description: The unique project ID
+              required: true
+              type: integer
+              default: 1
+        responses:
+            200:
+                description: Project Unfeatured
+            400:
+                description: Bad request
+            403:
+                description: Forbidden
+            404:
+                description: Project not found
+            500:
+                description: Internal Server Error
+        """
+        try:
+            ProjectService.unset_project_as_featured(project_id)
+            return {"Success": True}, 200
+        except NotFound:
+            return {"Error": "Project Not Found"}, 404
+        except ValueError as e:
+            error_msg = f"FeaturedProjects DELETE: {str(e)}"
+            return {"error": error_msg}, 400
+        except Exception as e:
+            error_msg = f"FeaturedProjects DELETE - unhandled error: {str(e)}"
+            current_app.logger.critical(error_msg)
+            return {"error": error_msg}, 500
